+++ Autotesting/app/src/androidTest/java/esadrcanfer/us/alumno/autotesting/inagraph/actions/RadioButtonInputGenerator.java
@@ -10,8 +10,8 @@ import androidx.test.uiautomator.UiSelector;
 
 public class RadioButtonInputGenerator extends InputGenerator {
 
-    Integer seed;
-    public RadioButtonInputGenerator(Integer seed){
+    Long seed;
+    public RadioButtonInputGenerator(Long seed){
         this.seed = seed;
     }
